requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg[binary]
gunicorn==21.2.0
whitenoise==6.6.0